refactor: Extract interface for dag.Store,dag.Read, and dag.Write
