+++ src/dag/mod.ts
@@ -1,5 +1,3 @@
-export {Read} from './read';
-export {Write} from './write';
 export type {Chunk, CreateChunk} from './chunk';
 export {
   createChunk,
@@ -7,6 +5,8 @@ export {
   createChunkWithHash,
   createChunkWithNativeHash,
 } from './chunk';
-export {Store} from './store';
+export type {Store, Read, Write} from './store';
+export {isWrite} from './store';
+export {StoreImpl} from './store-impl';
 export {TestStore} from './test-store';
 export * from './key';
